Move Document, Word to datatypes package, unit tests for Word.
