+++ src/main/java/nlmt/datatypes/SparseDocument.java
@@ -14,9 +14,7 @@
  * limitations under the License.
  */
 
-package nlmt.topicmodels;
-
-import nlmt.datatypes.IdentifierObjectMapper;
+package nlmt.datatypes;
 
 import java.util.*;
 import java.util.stream.Collectors;
