+++ src/test/java/nlmt/datatypes/DocumentTest.java
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nlmt.topicmodels;
+package nlmt.datatypes;
 
-import nlmt.datatypes.IdentifierObjectMapper;
 import org.junit.Before;
 import org.junit.Test;
