+++ src/test/java/nlmt/topicmodels/HierarchicalLDANodeTest.java
@@ -17,6 +17,7 @@
 package nlmt.topicmodels;
 
 import nlmt.datatypes.IdentifierObjectMapper;
+import nlmt.datatypes.Word;
 import org.junit.Before;
 import org.junit.Test;
